+++ src/constants.ts
@@ -1,8 +1,1 @@
-export enum Lifetime {
-  Resolve = 'RESOLVE',
-  Transient = 'TRANSIENT',
-  Scoped = 'SCOPED',
-  Singleton = 'SINGLETON',
-}
-
 export const DesignDependenciesKey = 'design:dependencies';