+++ src/enums.ts
@@ -1,0 +1,14 @@
+export enum Lifetime {
+  Resolve = 'RESOLVE',
+  Transient = 'TRANSIENT',
+  Scoped = 'SCOPED',
+  Singleton = 'SINGLETON',
+}
+
+export enum LogLevel {
+  Debug = 0,
+  Info = 1,
+  Warn = 2,
+  Error = 3,
+  None = 4,
+}